Move the progress task to its own module
